+++ source/patterns/@aws-solutions-constructs/aws-lambda-sqs/lib/index.ts
@@ -98,7 +98,7 @@ export class LambdaToSqs extends Construct {
 
         // Setup a dead letter queue, if applicable
         if (props.deployDeadLetterQueue || props.deployDeadLetterQueue === undefined) {
-          const dlq: sqs.Queue = defaults.buildQueue(this, 'deadLetterQueue', {
+          const [dlq] = defaults.buildQueue(this, 'deadLetterQueue', {
               queueProps: props.deadLetterQueueProps
           });
           this.deadLetterQueue = defaults.buildDeadLetterQueue({
@@ -108,7 +108,7 @@ export class LambdaToSqs extends Construct {
         }
 
         // Setup the queue
-        this.sqsQueue = defaults.buildQueue(this, 'queue', {
+        [this.sqsQueue] = defaults.buildQueue(this, 'queue', {
             existingQueueObj: props.existingQueueObj,
             queueProps: props.queueProps,
             deadLetterQueue: this.deadLetterQueue